backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
supabase==2.3.0
spacy==3.7.2
stanza==1.7.0
pytesseract==0.3.10
pdf2image==1.16.3
PyPDF2==3.0.1
python-docx==1.1.0
langdetect==1.0.9
sentence-transformers==2.2.2
Pillow==10.1.0
scikit-learn==1.3.2
numpy==1.26.2
